WIP: getting  random pictures to transition into top slide picture show. Also, drastically modifying tests to be more sound.

+++ Projects/LikeMyDessert.Repositories/Common/RepositoryBase.cs
@@ -22,7 +22,7 @@ namespace LikeMyDessert.Repositories
             UnitOfWork = unitOfWork;
         }
 
-        public long GetCount()
+        public int GetCount()
         {
             return UnitOfWork.GetCount<TPersistentObject>();
         }
